WIP: changed dbopts logic to use GORM
By having GORM integrated inside dbopts we can remove all the plain SQL
statements in the code and also have a generic way of creating and
querying our data that is independent of the database we want to use.

This also removes code complexity for new migrations and makes our
database operations more efficient.

Authored-by: Ignacio Elizaga <ielizaga@pivotal.io>